Upgrade Ember and replace Browserify with ember-cli-node-assets
This commit upgrade Ember and Ember CLI to 2.12. Bower dependency is
removed. Node 0.x support is dropped. Many dependencies are upgraded.

This commit also replaces ember-browserify with ember-cli-node-assets
as the solution of importing webcam.js and webcam.swf. Previous solution
requires ember-browserify to be added to the host application which is
not desired. This change may fix #6 as well.